# backend/openapi.yaml
openapi: 3.0.3
info:
  title: CRM API
  version: 1.0.0
  description: REST API for CRM system (users, customers, interactions, leads, tasks, reports)
servers:
  - url: http://localhost:4000
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Lead:
      type: object
      properties:
        id: { type: integer }
        customer_id: { type: integer }
        category_id: { type: integer }
        stage: { type: string }
        value: { type: number }
        source: { type: string }
    Task:
      type: object
      properties:
        id: { type: integer }
        user_id: { type: integer }
        customer_id: { type: integer }
        lead_id: { type: integer }
        category_id: { type: integer }
        title: { type: string }
        description: { type: string }
        due_date: { type: string, format: date-time }
        completed: { type: boolean }
...
security:
  - bearerAuth: []
paths:
...
  /api/leads:
    get:
      summary: List leads
      responses:
        '200': { description: List of leads }
    post:
      summary: Create lead
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Lead'
      responses:
        '201': { description: Lead created }
  /api/leads/{id}:
    get:
      summary: Get lead by ID
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        '200': { description: Lead found }
    put:
      summary: Update lead
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Lead'
      responses:
        '200': { description: Lead updated }
    delete:
      summary: Delete lead
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        '200': { description: Lead deleted }
  /api/tasks:
    get:
      summary: List tasks
      responses:
        '200': { description: List of tasks }
    post:
      summary: Create task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Task'
      responses:
        '201': { description: Task created }
  /api/tasks/{id}:
    get:
      summary: Get task by ID
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        '200': { description: Task found }
    put:
      summary: Update task
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Task'
      responses:
        '200': { description: Task updated }
    delete:
      summary: Delete task
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        '200': { description: Task deleted }
...